+++ runtime/src/index.ts
@@ -12,4 +12,11 @@ export { useConfig } from '@dhis2/app-service-config'
 
 export { useAlerts, useAlert } from '@dhis2/app-service-alerts'
 
+export {
+    useOnlineStatus,
+    useCacheableSection,
+    CacheableSection,
+    useCachedSections,
+} from '@dhis2/app-service-offline'
+
 export { Provider } from './Provider'